third_party/blink/web_tests/http/tests/security/subresourceIntegrity/subresource-integrity-style-allowed.html

<!DOCTYPE html>
<html>
    <head>
        <title>Subresource integrity allowed style resources</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <link rel="stylesheet" href="style-1-of-3.css" integrity="sha256-RvLeYLQyPa_ZQk95Rj0XQpfsoBHW9Vrqb3zwo5DScrI="></link>
        <link rel="stylesheet" href="style-2-of-3.css" integrity="sha384-j1bGpDxu0bZAb4jBJzciDWuVUuOTXlLvB9WNKDh7Gik9ygF6CctjgYISUwShHmFM"></link>
        <link rel="stylesheet" href="style-3-of-3.css" integrity="sha512-P302RfA5c_H9Ao_miirSIPVBr3uxOdTzm8IzaGwidneeOOwuyTfOhCp8-c1yg-W16rKf1IW_-4TxD0yAKdkV-Q=="></link>
    </head>
    <body>
        <p id="p1">Test 1</p>
        <p id="p2">Test 2</p>
        <p id="p3">Test 3</p>

        <script>
        setup({single_test: true});
        var color = window.getComputedStyle(document.querySelector('#p1')).color;
        assert_equals(color, "rgb(0, 128, 0)", "The p#1 element's text should be green.");
        color = window.getComputedStyle(document.querySelector('#p2')).color;
        assert_equals(color, "rgb(0, 128, 0)", "The p#2 element's text should be green.");
        color = window.getComputedStyle(document.querySelector('#p3')).color;
        assert_equals(color, "rgb(0, 128, 0)", "The p#3 element's text should be green.");
        done();
        </script>
    </body>
</html>
